+++ app/src/main/java/dev/jatzuk/mvvmrunning/other/TargetType.kt
@@ -1,5 +1,5 @@
 package dev.jatzuk.mvvmrunning.other
 
-enum class TargetType(var value: Long = 0) {
+enum class TargetType(var value: Float = 0f) {
     TIME, DISTANCE, CALORIES, NONE
 }
